move more logic into actions
