+++ pkg/core/actions/actions_test.go
@@ -106,17 +106,17 @@ func TestActionsImplementInterface(t *testing.T) {
 	processAction(&AllPods{})
 	processAction(&GetNextImage{})
 	processAction(&FinishScanClient{})
-	processAction(&GetNextImageForHubPolling{})
+	processAction(&GetInitialHubCheckImage{})
+	processAction(&InitialHubCheckResults{})
 	processAction(&HubCheckResults{})
-	processAction(&HubScanResults{})
 	processAction(&RequeueStalledScan{})
 	processAction(&SetConcurrentScanLimit{})
 	processAction(&AllImages{})
 	processAction(&GetModel{})
 	processAction(&GetMetrics{})
 	processAction(&GetScanResults{})
-	processAction(&GetInProgressHubScans{})
-	processAction(&GetInProgressScanClientScans{})
+	processAction(&GetRunningHubScans{})
+	processAction(&GetRunningScanClientScans{})
 	processAction(&HubRecheckResults{})
 	processAction(&GetCompletedScans{})
 }
