--- pkg/core/actions/getinprogressscanclientscans.go
@@ -1,40 +1,0 @@
-/*
-Copyright (C) 2018 Synopsys, Inc.
-
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements. See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership. The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License. You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied. See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
-
-package core
-
-import (
-	m "github.com/blackducksoftware/perceptor/pkg/core/model"
-)
-
-type GetInProgressScanClientScans struct {
-	Continuation func(imageInfos []*m.ImageInfo)
-}
-
-func (g *GetInProgressScanClientScans) Apply(model *m.Model) {
-	imageInfos := []*m.ImageInfo{}
-	for _, imageInfo := range model.InProgressScanClientScans() {
-		// TODO could make a deep copy of imageInfo in case it is being
-		// changed while we're looking at it
-		imageInfos = append(imageInfos, imageInfo)
-	}
-	go g.Continuation(imageInfos)
-}
